requirements.txt
torch==2.4.1
numpy==1.26.4
transformers==4.49.0.dev0
huggingface-hub==0.24.7
openai==1.44.0
google-generativeai==0.8.3
pandas
scipy
matplotlib
Pillow
pyyaml
loguru
tqdm
scikit-learn
affiliation
qwen-vl-utils
